Update README.md
Azure Site Recovery LAB
Scenario: Assume that I am in one of the organizations responsible for the infrastructure of Azure. I give the administrator permission (Microsoft Entra ID) to one of the employees responsible for (the database) to access various resources on Azure.

The latter created a virtual machine for me to work on (to manage the MySQl server database connection), then he wrote incorrect SQL queries and executed them, leading to unexpected changes or damaging the integrity of the data on the virtual machine, or I upgraded something on the virtual machine, which led to damage.  Pre-existing files and folders.

 In this case, you may need to completely restore the default device and restore it to its location before the problem.

To fix such errors  i use Azure Site Recovery, you can replicate virtual machines (VMs) from your primary data center or on-premises environment to Azure. Replication can be configured for both Hyper-V and VMware virtual machines, allowing flexibility in the source environment.